==> 2024-Ranek_etal_QUICHE/scripts_TNBC_preprocessing/10_collagen_fiber_segmentation_Spain.ipynb <==
import os
import matplotlib.pyplot as plt
import ark.segmentation.fiber_segmentation as fibseg

## Perform segmentation of collagen fibers

cohort_ID = 'SPAIN'
base_dir = f'/Volumes/Shared/Noah Greenwald/TNBC_cohorts/{cohort_ID}'

input_dir = os.path.join(base_dir, 'image_data', 'samples')
fibseg_dir = os.path.join(base_dir, 'data', 'fiber_segmentation_processed_data')

if not os.path.exists(fibseg_dir):
    os.makedirs(fibseg_dir)

example_fov_name = 'TMA31_R2C5'
fiber_channel_name = 'Collagen1'

blur = 2
CSD = 256
fiber_widths = (1, 3, 5, 7, 9)
ridge_cutoff = 0.1
sobel_blur = 1
min_fiber_size = 100

plt.style.use('dark_background')
fibseg.plot_fiber_segmentation_steps(input_dir, example_fov_name, fiber_channel_name,
    img_sub_folder=None,
    blur=blur,
    contrast_scaling_divisor=CSD,
    fiber_widths=fiber_widths,
    ridge_cutoff=ridge_cutoff,
    sobel_blur=sobel_blur,
    min_fiber_size=min_fiber_size
)
plt.gcf().set_size_inches((32, 32))

csv_compression = None

fiber_table = fibseg.run_fiber_segmentation(input_dir, fiber_channel_name, fibseg_dir,
    img_sub_folder=None,
    csv_compression=csv_compression,
    blur=blur,
    contrast_scaling_divisor=CSD,
    fiber_widths=fiber_widths,
    ridge_cutoff=ridge_cutoff,
    sobel_blur=sobel_blur,
    min_fiber_size=min_fiber_size,
    debug=False,
)

tile_length = 512
min_fiber_num = 5
save_tiles = True

fov_stats, tile_states = fibseg.generate_summary_stats(
    fiber_table, fibseg_dir, tile_length=tile_length, min_fiber_num=min_fiber_num, save_tiles=save_tiles)
